Avoid synapse's M_TOO_LARGE by chunking output
The approach is driven by three things:
 - Synapse's message json size limit is 64k
 - it's best to split messages on newlines
 - splitting in the middle of a utf8 character is bad

So, if we buffer more than 16k of shell output, we trigger an
'unflushed' send.

Next, when we send this way, we ignore the current (unterminated) line.
Unless we get 16K+ all without a newline, in which case we just send it
all.

Finally, we use a buffered incremental utf8 decoder to ensure we don't
try to decode a partial utf8 character.